requirements.txt
pandas
numpy
python-dotenv
requests
lyricsgenius
tqdm
matplotlib
spacy
sentence-transformers
umap-learn
plotly
scikit-learn
streamlit